+++ test/api/bulk_download_controller_test.rb
@@ -2,6 +2,7 @@ require 'api_test_helper'
 require 'test_helper'
 require 'user_tokens_helper'
 require 'bulk_download_helper'
+require 'includes_helper'
 
 class BulkDownloadControllerTest < ActionDispatch::IntegrationTest
