+++ test/api/metadata_schemas_controller_test.rb
@@ -1,5 +1,6 @@
 require 'api_test_helper'
 require 'test_helper'
+require 'includes_helper'
 
 class MetadataSchemasControllerTest < ActionDispatch::IntegrationTest
   include Api::V1::Concerns::ConventionSchemas
