+++ test/api/visualization/annotations_controller_test.rb
@@ -1,5 +1,6 @@
 require 'test_helper'
 require 'api_test_helper'
+require 'includes_helper'
 
 class AnnotationsControllerTest < ActionDispatch::IntegrationTest
