+++ test/controllers/analysis_configurations_controller_test.rb
@@ -1,6 +1,7 @@
 require 'integration_test_helper'
 require 'api_test_helper'
 require 'test_helper'
+require 'includes_helper'
 
 class AnalysisConfigurationsControllerTest < ActionDispatch::IntegrationTest
