+++ test/controllers/feature_flag_options_controller_test.rb
@@ -1,5 +1,6 @@
 require 'test_helper'
 require 'integration_test_helper'
+require 'includes_helper'
 
 class FeatureFlagOptionsControllerTest < ActionDispatch::IntegrationTest
