+++ test/controllers/site_controller_test.rb
@@ -1,5 +1,6 @@
 require 'integration_test_helper'
 require 'test_helper'
+require 'includes_helper'
 
 class SiteControllerTest < ActionDispatch::IntegrationTest
