+++ test/includes_helper.rb
@@ -1,0 +1,15 @@
+class ActionDispatch::IntegrationTest
+  include ::Minitest::Hooks
+  include ::SelfCleaningSuite
+  include ::TestInstrumentor
+  include ::Devise::Test::IntegrationHelpers
+  include ::Requests::JsonHelpers
+  include ::Requests::HttpHelpers
+end
+
+class ActionController::TestCase
+  include ::Minitest::Hooks
+  include ::SelfCleaningSuite
+  include ::TestInstrumentor
+  include ::Devise::Test::ControllerHelpers
+end
