+++ test/integration/download_agreement_test.rb
@@ -1,6 +1,6 @@
-require 'integration_test_helper'
 require 'api_test_helper'
 require 'test_helper'
+require 'includes_helper'
 
 class DownloadAgreementTest < ActionDispatch::IntegrationTest
