+++ test/integration/lib/search_facet_populator_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require 'includes_helper'
 
 class SearchFacetPopulatorTest < ActionDispatch::IntegrationTest
