+++ test/integration/synthetic_study_populator_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require 'includes_helper'
 
 class SyntheticStudyPopulatorTest < ActionDispatch::IntegrationTest
