+++ test/integration/upload_wizard_test.rb
@@ -1,5 +1,6 @@
 require 'integration_test_helper'
 require 'test_helper'
+require 'includes_helper'
 
 class UploadWizardTest < ActionDispatch::IntegrationTest
