+++ test/test_helper.rb
@@ -100,20 +100,3 @@ class ActiveSupport::TestCase
     end
   end
 end
-
-# include necessary modules in each test class
-class ActionDispatch::IntegrationTest
-  include ::Minitest::Hooks
-  include ::SelfCleaningSuite
-  include ::TestInstrumentor
-  include Devise::Test::IntegrationHelpers
-  include Requests::JsonHelpers
-  include Requests::HttpHelpers
-end
-
-class ActionController::TestCase
-  include ::Minitest::Hooks
-  include ::SelfCleaningSuite
-  include ::TestInstrumentor
-  include Devise::Test::ControllerHelpers
-end
